Extract getDeriveAccount, getDeriveAccountV3 function
